+++ lib/graphql/fragment_cache/cacher.rb
@@ -10,7 +10,26 @@ module GraphQL
         def call(query)
           return unless query.context.fragments?
 
-          query.context.fragments.each(&:persist)
+          if FragmentCache.cache_store.respond_to?(:write_multi)
+            batched_persist(query)
+          else
+            persist(query)
+          end
+        end
+
+        private
+
+        def batched_persist(query)
+          query.context.fragments.group_by(&:options).each do |options, group|
+            hash = group.map { |fragment| [fragment.cache_key, fragment.value] }.to_h
+            FragmentCache.cache_store.write_multi(hash, **options)
+          end
+        end
+
+        def persist(query)
+          query.context.fragments.each do |fragment|
+            FragmentCache.cache_store.write(fragment.cache_key, fragment.value, **fragment.options)
+          end
         end
       end
     end